Only allow one type of topo explorer per vtctld instance.
For migrations, just launch a separate vtctld for each topo type.

Also pull some code out from the monolithic vtctld.go for better
readability.
