+++ go/cmd/vtctld/explorer.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"errors"
+	"fmt"
 	"net/http"
 	"path"
 	"strings"
@@ -10,8 +12,6 @@ import (
 	"github.com/youtube/vitess/go/cmd/vtctld/proto"
 	"github.com/youtube/vitess/go/vt/topo"
 )
-
-var explorers = make(map[string]Explorer)
 
 // Explorer allows exploring a topology server.
 type Explorer interface {
@@ -52,16 +52,29 @@ type Explorer interface {
 	GetReplicationSlaves(cell, keyspace, shard string) string
 }
 
-var handleExplorerMutex sync.Mutex
+var (
+	// explorerMutex protects against concurrent registration attempts (e.g. OnRun).
+	// Other than registration, the explorer should never change.
+	explorerMutex sync.Mutex
+	explorerName  string
+	explorer      Explorer
+)
 
-// HandleExplorer serves explorer under url, using templateName.
-// It can be called concurrently, such as from servenv.OnRun hooks.
-func HandleExplorer(name, url, templateName string, explorer Explorer) {
-	handleExplorerMutex.Lock()
-	defer handleExplorerMutex.Unlock()
+// HandleExplorer registers the Explorer under url, using the given template.
+// It should be called by a plugin either from init() or from servenv.OnRun().
+// Only one Explorer can be registered in a given instance of vtctld.
+func HandleExplorer(name, url, templateName string, exp Explorer) {
+	explorerMutex.Lock()
+	defer explorerMutex.Unlock()
 
-	explorers[name] = explorer
+	if explorer != nil {
+		panic(fmt.Sprintf("Only one Explorer can be registered in vtctld. Trying to register %q, but %q was already registered.", name, explorerName))
+	}
+
+	explorer = exp
+	explorerName = name
 	indexContent.ToplevelLinks[name+" explorer"] = url
+
 	http.HandleFunc(url, func(w http.ResponseWriter, r *http.Request) {
 		if err := r.ParseForm(); err != nil {
 			httpError(w, "cannot parse form: %s", err)
@@ -80,4 +93,57 @@ func HandleExplorer(name, url, templateName string, explorer Explorer) {
 		result := explorer.HandlePath(actionRepo, topoPath, r)
 		templateLoader.ServeTemplate(templateName, result, w, r)
 	})
+}
+
+// handleExplorerRedirect returns the redirect target URL.
+func handleExplorerRedirect(r *http.Request) (string, error) {
+	keyspace := r.FormValue("keyspace")
+	shard := r.FormValue("shard")
+	cell := r.FormValue("cell")
+
+	switch r.FormValue("type") {
+	case "keyspace":
+		if keyspace == "" {
+			return "", errors.New("keyspace is required for this redirect")
+		}
+		return explorer.GetKeyspacePath(keyspace), nil
+	case "shard":
+		if keyspace == "" || shard == "" {
+			return "", errors.New("keyspace and shard are required for this redirect")
+		}
+		return explorer.GetShardPath(keyspace, shard), nil
+	case "srv_keyspace":
+		if keyspace == "" || cell == "" {
+			return "", errors.New("keyspace and cell are required for this redirect")
+		}
+		return explorer.GetSrvKeyspacePath(cell, keyspace), nil
+	case "srv_shard":
+		if keyspace == "" || shard == "" || cell == "" {
+			return "", errors.New("keyspace, shard, and cell are required for this redirect")
+		}
+		return explorer.GetSrvShardPath(cell, keyspace, shard), nil
+	case "srv_type":
+		tabletType := r.FormValue("tablet_type")
+		if keyspace == "" || shard == "" || cell == "" || tabletType == "" {
+			return "", errors.New("keyspace, shard, cell, and tablet_type are required for this redirect")
+		}
+		return explorer.GetSrvTypePath(cell, keyspace, shard, topo.TabletType(tabletType)), nil
+	case "tablet":
+		alias := r.FormValue("alias")
+		if alias == "" {
+			return "", errors.New("alias is required for this redirect")
+		}
+		tabletAlias, err := topo.ParseTabletAliasString(alias)
+		if err != nil {
+			return "", fmt.Errorf("bad tablet alias %q: %v", alias, err)
+		}
+		return explorer.GetTabletPath(tabletAlias), nil
+	case "replication":
+		if keyspace == "" || shard == "" || cell == "" {
+			return "", errors.New("keyspace, shard, and cell are required for this redirect")
+		}
+		return explorer.GetReplicationSlaves(cell, keyspace, shard), nil
+	default:
+		return "", errors.New("bad redirect type")
+	}
 }
